+++ demo/symfony8/importmap.php
@@ -49,4 +49,11 @@ return [
         'version' => '2.5.2',
         'type'    => 'css',
     ],
+    '@symfony/ux-live-component' => [
+        'path' => './vendor/symfony/ux-live-component/assets/dist/live_controller.js',
+    ],
+    '@symfony/ux-live-component/dist/live.min.css' => [
+        'path' => './vendor/symfony/ux-live-component/assets/dist/live.min.css',
+        'type' => 'css',
+    ],
 ];